pyproject.toml
[project]
name = "pgt"
version = "0.1.0"
description = "PostgreSQL/Python application with intelligent search, ingestion pipeline, and AI agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Database
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "sqlalchemy[asyncio]>=2.0.0",
    # Embeddings & AI
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "tiktoken>=0.7.0",
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    # CLI & Output
    "rich>=13.0.0",
    "typer>=0.12.0",
    # Async & Queue
    "aiofiles>=24.0.0",
    "aiohttp>=3.9.0",
    # Configuration
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=6.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
